Provide an easy way to run acceptance test via rake rspec
When either ZABBIX_RUN_ACCEPTANCE or ZABBIX_HOST_URL is set as an
environment variable, it's considered that the user opts in to
acceptance tests.

The acceptance tests previously didn't have the _spec suffix, so they
weren't picked up by the file filter. The acceptance tests are dangerous
on a production system, so they are now flagged with their own type.
Flagging also makes it possible to include specific contexts just for
acceptance tests.